Collapsing the end exam actions to a single url, cleaning stale comment
