+++ apps/prairielearn/src/pages/endExam/endExam.ts
@@ -3,17 +3,19 @@ import asyncHandler from 'express-async-handler';
 import { z } from 'zod';
 
 import { HttpStatusError } from '@prairielearn/error';
+import { html } from '@prairielearn/html';
 import { logger } from '@prairielearn/logger';
 import { loadSqlEquiv, queryOptionalRow } from '@prairielearn/postgres';
 
 import { generateEndExamJwt } from '../../ee/auth/endExamJwt.js';
-import { getEndExamCloseUrl } from '../../lib/client/url.js';
+import { getEndExamCloseUrl, getEndExamExitUrl } from '../../lib/client/url.js';
 import { config } from '../../lib/config.js';
 
 const router = Router();
 const sql = loadSqlEquiv(import.meta.url);
 
 const PT_END_EXAM_TIMEOUT_MS = 10_000;
+const CLOSE_URL = getEndExamExitUrl();
 
 /**
  * Finds the user's currently-active LDB-required reservation, looked up at
@@ -35,6 +37,62 @@ export async function selectActiveLockdownBrowserReservation({
     z.object({ id: z.string() }),
   );
 }
+
+/**
+ * PL-hosted two-hop LDB exit handshake. The POST handler redirects back to this
+ * route with `?rldbsm=1` (LDB drops to medium security); the page then
+ * meta-refreshes to `?rldbxb=1`, which LDB intercepts to exit.
+ *
+ * Falls back to a static page when not in an LDB session, when the close
+ * handshake was not requested, or when `rldbxb=1` reaches the server (LDB
+ * should have intercepted it), to avoid an infinite refresh loop.
+ */
+router.get(
+  '/',
+  asyncHandler(async (req, res) => {
+    // Presence check, not equality — Express's qs parser can yield an
+    // array for duplicated params.
+    if (!req.session.lockdown_browser || !('rldbsm' in req.query) || 'rldbxb' in req.query) {
+      res.send(
+        html`<!doctype html>
+          <html lang="en">
+            <head>
+              <meta charset="utf-8" />
+              <title>LockDown Browser closed</title>
+              <meta name="robots" content="noindex" />
+            </head>
+            <body>
+              <main>
+                <h1>LockDown Browser closed</h1>
+                <p>You may now close this window.</p>
+              </main>
+            </body>
+          </html>`.toString(),
+      );
+      return;
+    }
+
+    res.send(
+      html`<!doctype html>
+        <html lang="en">
+          <head>
+            <meta charset="utf-8" />
+            <title>Closing LockDown Browser…</title>
+            <meta name="robots" content="noindex" />
+            <meta http-equiv="refresh" content="0;url=${CLOSE_URL}" />
+          </head>
+          <body>
+            <main>
+              <h1>Closing LockDown Browser…</h1>
+              <p>
+                <a href="${CLOSE_URL}">Click here if the browser does not close automatically.</a>
+              </p>
+            </main>
+          </body>
+        </html>`.toString(),
+    );
+  }),
+);
 
 /**
  * Receives the click on the LDB navbar End exam button: looks up the
